fix: teste unitarios service

+++ src/main/java/br/com/template/model/dto/UsuarioDTO.java
@@ -1,6 +1,6 @@
 package br.com.template.model.dto;
 
-public record UsuarioDTO(Long id,
+public record UsuarioDTO(Integer id,
                          String email,
                          Integer idstatus) {
 }
